+++ SSDAM/CHANGE_LOG.md
@@ -1,7 +1,12 @@
-## Memory Manager (v1.0.1)
+## SSDAM (v1.1.0)
+* 8.2x times faster than traditional new/delete.
+* const non-pointer/non-reference class members are allowed (MemoryManagerConsBased.cpp). Speed will be 7.43x not 8.2x.
+* const non-pointer/non-reference class members are allowed (MemoryManagerNewDelete.cpp). Speed will be 6.78x not 8.2x.
+
+## SSDAM (v1.0.1)
 * 8x times faster than traditional new/delete.
 
-## Memory Manager (v1.0.0)
+## SSDAM (v1.0.0)
 * memory safe
 * single threaded
 * 6x times faster than traditional new/delete.